% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{CalculateTrainPredictYearPattern}
\alias{CalculateTrainPredictYearPattern}
\title{PROJ}
\usage{
CalculateTrainPredictYearPattern(yearMin, yearMax, numPerYear1 = 1)
}
\arguments{
\item{yearMin}{a}

\item{yearMax}{a}

\item{numPerYear1}{a}
}
\description{
PROJ
}
