% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckForOutbreaksUtbrudd.r
\name{CheckForOutbreaksUtbrudd}
\alias{CheckForOutbreaksUtbrudd}
\title{test}
\usage{

  CheckForOutbreaksUtbrudd(resYearLine = readRDS(fhi::DashboardFolder("results",
  "resYearLine.RDS")))
}
\description{
test
}
