% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CreateLatestDoneFile}
\alias{CreateLatestDoneFile}
\title{test}
\usage{
CreateLatestDoneFile(file = fhi::DashboardFolder("data_clean", paste0("done_",
  LatestRawID(), ".txt")))
}
\arguments{
\item{file}{a}
}
\description{
test
}
