% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailInternal}
\alias{EmailInternal}
\title{test}
\usage{
EmailInternal(resYearLine = readRDS(fhi::DashboardFolder("results",
  "resYearLine.RDS")), isTest = TRUE)
}
\arguments{
\item{resYearLine}{a}

\item{isTest}{a}
}
\description{
test
}
