% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationOfNewData}
\alias{EmailNotificationOfNewData}
\title{test}
\usage{
EmailNotificationOfNewData(files, isTest = TRUE)
}
\arguments{
\item{files}{a}

\item{isTest}{a}
}
\description{
test
}
