% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationOfNewResults}
\alias{EmailNotificationOfNewResults}
\title{test}
\usage{

  EmailNotificationOfNewResults(lastEmailedUtbruddFile = fhi::DashboardFolder("results",
  "lastEmailedUtbrudd.RDS"))
}
\arguments{
\item{lastEmailedUtbruddFile}{a}
}
\description{
test
}
