% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationUtbrudd}
\alias{EmailNotificationUtbrudd}
\title{test}
\usage{

  EmailNotificationUtbrudd(lastEmailedUtbruddFile = fhi::DashboardFolder("results",
  "lastEmailedUtbrudd.RDS"))
}
\arguments{
\item{lastEmailedUtbruddFile}{a}
}
\description{
test
}
