% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington.R
\name{FarringtonSEinGammaSpace}
\alias{FarringtonSEinGammaSpace}
\title{test}
\usage{
FarringtonSEinGammaSpace(pred, phi, alpha = NULL, z = NULL,
  skewness.transform = "none")
}
\arguments{
\item{pred}{a}

\item{phi}{a}

\item{alpha}{a}

\item{z}{a}

\item{skewness.transform}{a}
}
\description{
test
}
