% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrington.R
\name{FarringtonZscore}
\alias{FarringtonZscore}
\title{test}
\usage{
FarringtonZscore(pred, phi, alpha = NULL, z = NULL,
  skewness.transform = "none", y)
}
\arguments{
\item{pred}{a}

\item{phi}{a}

\item{alpha}{a}

\item{z}{a}

\item{skewness.transform}{a}

\item{y}{a}
}
\description{
test
}
