% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{FormatData}
\alias{FormatData}
\title{test}
\usage{
FormatData(d, population = readRDS(fhi::DashboardFolder("data_clean",
  "pop.RDS")), hellidager = fread(system.file("extdata",
  "DatoerMedHelligdager.txt", package = "sykdomspuls"))[, c("Dato",
  "HelligdagIndikator"), with = FALSE],
  testIfHelligdagIndikatorFileIsOutdated = TRUE)
}
\arguments{
\item{d}{a}

\item{population}{a}

\item{hellidager}{a}

\item{testIfHelligdagIndikatorFileIsOutdated}{a}
}
\description{
test
}
