% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateData.R
\name{GenerateAnalysisResults}
\alias{GenerateAnalysisResults}
\title{GenerateAnalysisResults}
\usage{
GenerateAnalysisResults(granularity = "weekly", loc = "Norge",
  type = CONFIG$SYNDROMES_ALERT_EXTERNAL[1])
}
\arguments{
\item{granularity}{a}

\item{loc}{a}

\item{type}{a}
}
\description{
GenerateAnalysisResults
}
