% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListOutbreaks.R
\name{GenerateOutbreakListExternal}
\alias{GenerateOutbreakListExternal}
\title{test}
\usage{
GenerateOutbreakListExternal(df = readRDS(fhi::DashboardFolder("results",
  "resYearLine.RDS")), dk = readRDS(fhi::DashboardFolder("results",
  "resYearLineMunicip.RDS")), saveFiles = fhi::DashboardFolder("results",
  "outbreaks_alert_external.RDS"),
  alerts = readxl::read_excel(file.path("/etc", "gmailr",
  "emails_sykdomspuls_alert.xlsx")))
}
\arguments{
\item{df}{a}

\item{dk}{a}

\item{saveFiles}{a}

\item{alerts}{a}
}
\description{
test
}
