% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListOutbreaks.R
\name{GenerateOutbreakListInternal}
\alias{GenerateOutbreakListInternal}
\title{test}
\usage{
GenerateOutbreakListInternal(df = readRDS(fhi::DashboardFolder("results",
  "resYearLine.RDS")), dk = readRDS(fhi::DashboardFolder("results",
  "resYearLineMunicip.RDS")), saveFiles = c(fhi::DashboardFolder("results",
  "outbreaks.RDS"), fhi::DashboardFolder("data_app", "outbreaks.RDS")))
}
\arguments{
\item{df}{a}

\item{dk}{a}

\item{saveFiles}{a}
}
\description{
test
}
