% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{GetPopulation}
\alias{GetPopulation}
\title{test}
\usage{
GetPopulation(L = c(0, 5, 15, 20, 30, 65), U = c(4, 14, 19, 29, 64, 9999),
  saveFiles = fhi::DashboardFolder("data_clean", "pop.RDS"))
}
\arguments{
\item{L}{a}

\item{U}{a}

\item{saveFiles}{a}
}
\description{
test
}
