% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{IdentifyDatasets}
\alias{IdentifyDatasets}
\title{test}
\usage{
IdentifyDatasets(raw = list.files(fhi::DashboardFolder("data_raw"),
  "^partially_formatted_"),
  clean = list.files(fhi::DashboardFolder("data_clean"), "done_"))
}
\arguments{
\item{raw}{a}

\item{clean}{a}
}
\description{
test
}
