% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{LATEST_DATA}
\alias{LATEST_DATA}
\title{LATEST_DATA}
\format{An object of class \code{environment} of length 0.}
\usage{
LATEST_DATA
}
\description{
LATEST_DATA
}
\keyword{datasets}
