% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{norwayLocations}
\alias{norwayLocations}
\title{norwayLocations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 426 rows and 9 columns.}
\usage{
norwayLocations
}
\description{
norwayLocations
}
\keyword{datasets}
