% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{norwayMunicipMerging}
\alias{norwayMunicipMerging}
\title{norwayMunicipMerging}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5571 rows and 9 columns.}
\usage{
norwayMunicipMerging
}
\description{
norwayMunicipMerging
}
\keyword{datasets}
