% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasipoisson.algorithm.r
\name{quasipoisson.algorithm}
\alias{quasipoisson.algorithm}
\title{test}
\usage{
quasipoisson.algorithm(dataset, predinterval = 30,
  historical.data.years = 5, mod.pred.window = 90, reweights = 1,
  remove.pandemic.year = F, remove.highcounts = 0, sign.level = 0.05,
  v = 1)
}
\description{
test
}
