% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuasipoissonTestPredictData.r
\name{QuasipoissonTrainPredictData}
\alias{QuasipoissonTrainPredictData}
\title{test}
\usage{
QuasipoissonTrainPredictData(datasetTrain, datasetPredict, reweights = 1,
  remove.pandemic.year = F, remove.highcounts = 0, sign.level = 0.05,
  isDaily = TRUE, v = 1)
}
\arguments{
\item{datasetTrain}{a}

\item{datasetPredict}{a}

\item{reweights}{a}

\item{remove.pandemic.year}{a}

\item{remove.highcounts}{a}

\item{sign.level}{a}

\item{isDaily}{a}

\item{v}{a}
}
\description{
test
}
