% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{RunOneAnalysis}
\alias{RunOneAnalysis}
\title{RunOneAnalysis}
\usage{
RunOneAnalysis(analysesStack, analysisData)
}
\arguments{
\item{analysesStack}{a}

\item{analysisData}{a}
}
\description{
RunOneAnalysis
}
