% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateData.R
\name{ValidateAnalysisResults}
\alias{ValidateAnalysisResults}
\title{ValidateDataClean}
\usage{
ValidateAnalysisResults(d, granularity = "weekly")
}
\arguments{
\item{d}{a}

\item{granularity}{a}
}
\description{
ValidateDataClean
}
