% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesAlgorithmBasic}
\alias{variablesAlgorithmBasic}
\title{variablesAlgorithmBasic}
\format{An object of class \code{character} of length 4.}
\usage{
variablesAlgorithmBasic
}
\description{
variablesAlgorithmBasic
}
\keyword{datasets}
