% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesAlgorithmDaily}
\alias{variablesAlgorithmDaily}
\title{variablesAlgorithmDaily}
\format{An object of class \code{character} of length 1.}
\usage{
variablesAlgorithmDaily
}
\description{
variablesAlgorithmDaily
}
\keyword{datasets}
