% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesAlgorithmProduced}
\alias{variablesAlgorithmProduced}
\title{variablesAlgorithmProduced}
\format{An object of class \code{character} of length 9.}
\usage{
variablesAlgorithmProduced
}
\description{
variablesAlgorithmProduced
}
\keyword{datasets}
