% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesAlgorithmWeekly}
\alias{variablesAlgorithmWeekly}
\title{variablesAlgorithmWeekly}
\format{An object of class \code{character} of length 2.}
\usage{
variablesAlgorithmWeekly
}
\description{
variablesAlgorithmWeekly
}
\keyword{datasets}
