% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesMunicip}
\alias{variablesMunicip}
\title{variablesMunicip}
\format{An object of class \code{character} of length 1.}
\usage{
variablesMunicip
}
\description{
variablesMunicip
}
\keyword{datasets}
