% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesPostProcessing}
\alias{variablesPostProcessing}
\title{variablesPostProcessing}
\format{An object of class \code{character} of length 5.}
\usage{
variablesPostProcessing
}
\description{
variablesPostProcessing
}
\keyword{datasets}
