% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailAlertExternal}
\alias{EmailAlertExternal}
\title{test}
\usage{
EmailAlertExternal(results = readRDS(fhi::DashboardFolder("results",
  "outbreaks_alert_external.RDS")))
}
\arguments{
\item{results}{a}
}
\description{
test
}
